src/datagen/scenario-1/hgfhfg/requirements.txt
psycopg2-binary==2.9.1
Jinja2==3.0.3
PyYAML~=6.0.2
tqdm~=4.67.0
black~=24.10.0
pytz~=2024.2
graphviz